refactor(editor): extract events and setup slices (PR 10.8–10.14)
Finish Track B event peels and Track C setup owners without behavior changes:

- events: crop apply/reset, tool-mode switches, tool options
- setup: inspector family builders, async effects service
- setup: background assets + canvas layout services
- setup: window chrome + empty drop-zone reload path

Update refactor status/validation docs. Focused editor tests and PR gates green.
