+++ src/capture/editor/window/mod.rs
@@ -1,17 +1,15 @@
 use gdk4x11::X11Surface;
 use gtk4::gdk;
-use gtk4::gio;
 use gtk4::{
-    glib, prelude::*, Align, Application, ApplicationWindow, Box as GtkBox, Button, CheckButton,
-    DrawingArea, DropTarget, Entry, FileChooserAction, FileChooserNative, FileFilter, Image, Label,
-    Orientation, Overlay, Popover, ResponseType, Revealer, Spinner,
+    glib, prelude::*, Application, ApplicationWindow, Box as GtkBox, Button, CheckButton,
+    DrawingArea, DropTarget, Entry, Label, Orientation, Overlay, Popover,
 };
 use image::RgbaImage;
 use std::cell::{Cell, RefCell};
 use std::path::PathBuf;
 use std::rc::Rc;
-use std::sync::{mpsc, Arc, Mutex};
-use std::time::{Duration, SystemTime, UNIX_EPOCH};
+use std::sync::{Arc, Mutex};
+use std::time::{SystemTime, UNIX_EPOCH};
 use x11rb::{connection::Connection, protocol::xproto, protocol::xproto::ConnectionExt};
 
 use self::background_panel::BACKGROUND_SIDEBAR_WIDTH;
@@ -25,10 +23,10 @@ use super::render::{
     draw_text_edit_border, draw_text_edit_handles, rgba_image_to_surface, text_action_bounds,
 };
 use super::selection::{action_bounds_with_padding, action_resize_handles};
-use super::state::{apply_effect_actions, render_shadow_layer, EditorState};
+use super::state::{render_shadow_layer, EditorState};
 use super::types::{
-    tool_button_index, AnnotationAction, ArrowStyle, BackgroundAlignment, BackgroundStyle,
-    CropAspectRatio, DrawColor, EditorError, Point, Rect, Tool, ViewTransform,
+    tool_button_index, AnnotationAction, ArrowStyle, BackgroundStyle, CropAspectRatio, DrawColor,
+    EditorError, Point, Rect, Tool, ViewTransform,
 };
 
 const MAX_PREVIEW_SHADOW_DIM: u32 = 1200;
@@ -204,10 +202,9 @@ fn selected_action_geometry(action: &AnnotationAction) -> String {
 }
 
 use super::ui_support::{
-    arrow_style_toolbar_icon, install_edge_resize, install_editor_css, install_top_bar_window_drag,
-    install_window_drag, prefers_dark_glass_theme, prefers_reduced_transparency,
-    recommended_window_size_with_extra_width, set_active_tool_button, tool_icon_widget,
-    toolbar_icon_size, EDITOR_MIN_WINDOW_WIDTH, EDITOR_TOP_CHROME_HEIGHT,
+    arrow_style_toolbar_icon, install_editor_css, prefers_dark_glass_theme,
+    prefers_reduced_transparency, recommended_window_size_with_extra_width, set_active_tool_button,
+    tool_icon_widget, toolbar_icon_size, EDITOR_MIN_WINDOW_WIDTH, EDITOR_TOP_CHROME_HEIGHT,
 };
 
 const TEXT_SIZE_OPTIONS: [i32; 12] = [12, 14, 16, 18, 20, 24, 28, 32, 36, 48, 64, 72];
@@ -335,17 +332,23 @@ fn sync_obfuscate_option_selection(list: &GtkBox, selected_index: usize) {
     }
 }
 
+mod background_assets;
 pub mod background_panel;
 mod canvas;
+mod canvas_layout;
+mod chrome;
 pub mod color_picker;
 pub mod colors_panel;
 #[allow(dead_code)]
 mod cursor;
+mod effects;
+mod empty_state;
 mod events;
 mod footer;
 mod inspectors;
 mod toolbar;
 
+use background_assets::BackgroundAssetCaches;
 use inspectors::InspectorParts;
 
 #[allow(dead_code)]
@@ -386,138 +389,6 @@ pub fn open_image_editor_empty() -> Result<(), EditorError> {
 
     let _ = app.run_with_args::<String>(&[]);
     Ok(())
-}
-
-fn is_supported_image_path(path: &std::path::Path) -> bool {
-    matches!(
-        path.extension()
-            .and_then(|e| e.to_str())
-            .map(|e| e.to_lowercase())
-            .as_deref(),
-        Some("png") | Some("jpg") | Some("jpeg") | Some("webp")
-    )
-}
-
-/// Validate + decode the image on a background thread (showing the loading
-/// banner, like the video editor's async video load), then rebuild the
-/// editor in the SAME window with the file loaded.
-#[allow(clippy::too_many_arguments)]
-fn load_image_into_editor(
-    app: &Application,
-    window: &ApplicationWindow,
-    path: PathBuf,
-    loading_revealer: &Revealer,
-    loading_spinner: &Spinner,
-    loading: Rc<Cell<bool>>,
-    open_button: &Button,
-) {
-    if loading.get() {
-        return;
-    }
-    loading.set(true);
-    loading_revealer.set_reveal_child(true);
-    loading_spinner.set_visible(true);
-    loading_spinner.start();
-    open_button.set_sensitive(false);
-
-    let (sender, receiver) = mpsc::channel::<Result<(), String>>();
-    let path_thread = path.clone();
-    std::thread::spawn(move || {
-        let result = image::open(&path_thread)
-            .map(|_| ())
-            .map_err(|e| e.to_string());
-        let _ = sender.send(result);
-    });
-
-    let app = app.clone();
-    let window = window.clone();
-    let loading_revealer = loading_revealer.clone();
-    let loading_spinner = loading_spinner.clone();
-    let open_button = open_button.clone();
-    glib::timeout_add_local(Duration::from_millis(100), move || {
-        let stop_loading = || {
-            loading.set(false);
-            loading_revealer.set_reveal_child(false);
-            loading_spinner.stop();
-            loading_spinner.set_visible(false);
-            open_button.set_sensitive(true);
-        };
-        match receiver.try_recv() {
-            Ok(Ok(())) => {
-                stop_loading();
-                // Rebuild the editor UI with the image, reusing this window.
-                setup_editor_window_full(&app, path.clone(), false, Some(window.clone()));
-                glib::ControlFlow::Break
-            }
-            Ok(Err(err)) => {
-                stop_loading();
-                eprintln!("[editor] Failed to load image: {err}");
-                glib::ControlFlow::Break
-            }
-            Err(mpsc::TryRecvError::Empty) => glib::ControlFlow::Continue,
-            Err(mpsc::TryRecvError::Disconnected) => {
-                stop_loading();
-                glib::ControlFlow::Break
-            }
-        }
-    });
-}
-
-#[allow(clippy::too_many_arguments)]
-fn show_open_image_dialog(
-    app: &Application,
-    parent: &ApplicationWindow,
-    loading_revealer: &Revealer,
-    loading_spinner: &Spinner,
-    loading: Rc<Cell<bool>>,
-    open_button: &Button,
-) {
-    let chooser = FileChooserNative::new(
-        Some("Open image"),
-        Some(parent),
-        FileChooserAction::Open,
-        Some("Open"),
-        Some("Cancel"),
-    );
-
-    let filter = FileFilter::new();
-    filter.set_name(Some("Image files"));
-    filter.add_mime_type("image/png");
-    filter.add_mime_type("image/jpeg");
-    filter.add_mime_type("image/webp");
-    filter.add_pattern("*.png");
-    filter.add_pattern("*.jpg");
-    filter.add_pattern("*.jpeg");
-    filter.add_pattern("*.webp");
-    chooser.add_filter(&filter);
-
-    let app_ref = app.clone();
-    let parent_ref = parent.clone();
-    let revealer_ref = loading_revealer.clone();
-    let spinner_ref = loading_spinner.clone();
-    let open_button_ref = open_button.clone();
-    chooser.connect_response(move |dialog, response| {
-        if response == ResponseType::Accept {
-            if let Some(file) = dialog.file() {
-                if let Some(path) = file.path() {
-                    if is_supported_image_path(&path) {
-                        // Load into the same window, like the video editor.
-                        load_image_into_editor(
-                            &app_ref,
-                            &parent_ref,
-                            path.to_path_buf(),
-                            &revealer_ref,
-                            &spinner_ref,
-                            loading.clone(),
-                            &open_button_ref,
-                        );
-                    }
-                }
-            }
-        }
-        dialog.hide();
-    });
-    chooser.show();
 }
 
 #[allow(unused_imports)]
@@ -1397,213 +1268,14 @@ fn setup_editor_window_full(
     let cached_shadow_surface = Rc::new(std::cell::RefCell::new(None::<gtk4::cairo::ImageSurface>));
     let cached_shadow_signature = Rc::new(Cell::new(None::<(u32, u32, u64, u64, u64)>));
 
-    let gradient_surfaces = Rc::new(RefCell::new(vec![
-            None::<gtk4::cairo::ImageSurface>;
-            background_panel::BACKGROUND_GRADIENT_PREVIEW_FILES.len()
-        ]));
-    let wallpaper_cache = Rc::new(RefCell::new(std::collections::HashMap::<
-        PathBuf,
-        gtk4::cairo::ImageSurface,
-    >::new()));
+    let BackgroundAssetCaches {
+        gradient_surfaces,
+        wallpaper_cache,
+        wallpaper_loader_sender,
+    } = background_assets::install_background_asset_loading(&drawing_area);
 
-    let (wallpaper_loader_sender, receiver) =
-        std::sync::mpsc::channel::<(Option<usize>, PathBuf, RgbaImage)>();
-
-    // Pre-load gradients and system wallpaper in background
-    {
-        let sender = wallpaper_loader_sender.clone();
-        // Background loader thread
-        std::thread::spawn({
-            move || {
-                // 1. System wallpaper (High Priority)
-                if let Some(path) = background_panel::detect_system_wallpaper_path() {
-                    println!("[DEBUG] Detected system wallpaper: {:?}", path);
-                    if let Some(rgba) = background_panel::load_background_image_optimized(&path) {
-                        let _ = sender.send((None, path, rgba));
-                    }
-                } else {
-                    println!("[DEBUG] No system wallpaper detected.");
-                    // Also load the fallback wallpaper into cache
-                    let fallback_path = background_panel::background_gradient_asset_path(
-                        background_panel::BACKGROUND_GRADIENT_PREVIEW_FILES[0],
-                    );
-                    if let Some(rgba) =
-                        background_panel::load_background_image_optimized(&fallback_path)
-                    {
-                        let _ = sender.send((None, fallback_path, rgba));
-                    }
-                }
-
-                // 2. Gradients
-                for (idx, file_name) in background_panel::BACKGROUND_GRADIENT_PREVIEW_FILES
-                    .iter()
-                    .enumerate()
-                {
-                    let path = background_panel::background_gradient_asset_path(file_name);
-                    if let Some(rgba) = background_panel::load_background_image_optimized(&path) {
-                        if sender.send((Some(idx), path, rgba)).is_err() {
-                            break;
-                        }
-                    }
-                }
-            }
-        });
-
-        let gradient_surfaces_main = gradient_surfaces.clone();
-        let wallpaper_cache_main = wallpaper_cache.clone();
-        let drawing_area_main = drawing_area.downgrade();
-        glib::timeout_add_local(std::time::Duration::from_millis(100), move || {
-            while let Ok((idx_opt, path, rgba)) = receiver.try_recv() {
-                if let Some(surface) = rgba_image_to_surface(&rgba) {
-                    if let Some(idx) = idx_opt {
-                        gradient_surfaces_main.borrow_mut()[idx] = Some(surface);
-                    } else {
-                        wallpaper_cache_main.borrow_mut().insert(path, surface);
-                    }
-                    if let Some(area) = drawing_area_main.upgrade() {
-                        area.queue_draw();
-                    }
-                }
-            }
-            glib::ControlFlow::Continue
-        });
-    }
-
-    // Async Effects Pipeline
-    let (effects_sender, effects_receiver) = std::sync::mpsc::channel::<(RgbaImage, u64)>();
-    let (request_sender, request_receiver) =
-        std::sync::mpsc::channel::<(Arc<RgbaImage>, Vec<AnnotationAction>, u64)>();
-
-    // Used by the UI thread to coalesce effect rebuild requests.
-    let effects_request_sender = request_sender.clone();
-
-    let state_effects = state.clone();
-    let drawing_area_effects = drawing_area.downgrade();
-    {
-        glib::timeout_add_local(std::time::Duration::from_millis(16), move || {
-            while let Ok((new_image, revision)) = effects_receiver.try_recv() {
-                // Apply results, then if another rebuild was requested while pending,
-                // schedule one more rebuild.
-                let (should_schedule_next, base_image, actions, next_revision) = {
-                    let mut st = state_effects.lock().unwrap();
-                    if revision <= st.last_applied_effect_revision {
-                        (false, None, None, 0)
-                    } else {
-                        st.working_image = Arc::new(new_image);
-                        st.last_applied_effect_revision = revision;
-                        st.select_effect_rebuild_pending = false;
-                        st.mark_working_image_dirty();
-
-                        let should = st.select_effect_rebuild_dirty;
-                        if should {
-                            st.select_effect_rebuild_dirty = false;
-                            st.select_effect_rebuild_pending = true;
-                            st.pending_effect_revision += 1;
-                            (
-                                true,
-                                Some(Arc::clone(&st.base_image)),
-                                Some(st.actions.clone()),
-                                st.pending_effect_revision,
-                            )
-                        } else {
-                            (false, None, None, 0)
-                        }
-                    }
-                };
-
-                if let Some(area) = drawing_area_effects.upgrade() {
-                    area.queue_draw();
-                }
-
-                if should_schedule_next {
-                    if let (Some(base_image), Some(actions)) = (base_image, actions) {
-                        let _ = effects_request_sender.send((base_image, actions, next_revision));
-                    }
-                }
-            }
-            glib::ControlFlow::Continue
-        });
-    }
-
-    // Single background worker thread
-    std::thread::spawn(move || {
-        while let Ok(mut request) = request_receiver.recv() {
-            // Drain the channel to get only the latest request
-            while let Ok(newer) = request_receiver.try_recv() {
-                request = newer;
-            }
-
-            let (base_image, actions, revision) = request;
-            let mut working_image = (*base_image).clone();
-
-            // EXPENSIVE: This blocks the worker thread
-            apply_effect_actions(&mut working_image, &actions);
-
-            let _ = effects_sender.send((working_image, revision));
-        }
-    });
-
-    let rebuild_effects_async: Rc<dyn Fn()> = Rc::new({
-        let state = state.clone();
-        let sender = request_sender;
-        move || {
-            let maybe_payload = {
-                let mut st = state.lock().unwrap();
-
-                // Avoid flooding the worker with rebuild requests while one is already pending.
-                // This helps prevent UI stalls when many effect-triggering actions happen quickly.
-                if st.select_effect_rebuild_pending {
-                    // A rebuild is already in-flight; remember that we need another pass.
-                    st.select_effect_rebuild_dirty = true;
-                    return;
-                }
-                st.select_effect_rebuild_pending = true;
-                st.select_effect_rebuild_dirty = false;
-                st.last_effect_request_time_us = glib::monotonic_time();
-
-                st.pending_effect_revision += 1;
-                Some((
-                    Arc::clone(&st.base_image),
-                    st.actions.clone(),
-                    st.pending_effect_revision,
-                ))
-            };
-
-            if let Some((base_image, actions, revision)) = maybe_payload {
-                let _ = sender.send((base_image, actions, revision));
-            }
-        }
-    });
-
-    // Effects rebuild watchdog: if we ever get stuck with `select_effect_rebuild_pending=true`
-    // (e.g., app was backgrounded / main loop paused), recover by clearing pending and
-    // scheduling a fresh rebuild.
-    {
-        let state = state.clone();
-        let rebuild_effects_async = rebuild_effects_async.clone();
-        glib::timeout_add_local(std::time::Duration::from_millis(500), move || {
-            let should_recover = {
-                let st = state.lock().unwrap();
-                if !st.select_effect_rebuild_pending {
-                    false
-                } else {
-                    let elapsed = glib::monotonic_time() - st.last_effect_request_time_us;
-                    // 2 seconds without a result is considered stuck.
-                    elapsed > 2_000_000
-                }
-            };
-
-            if should_recover {
-                {
-                    let mut st = state.lock().unwrap();
-                    st.select_effect_rebuild_pending = false;
-                }
-                rebuild_effects_async();
-            }
-
-            glib::ControlFlow::Continue
-        });
-    }
+    // Async Effects Pipeline (channels, worker, polling, watchdog, rebuild callback).
+    let rebuild_effects_async = effects::install_async_effects_pipeline(&state, &drawing_area);
 
     let background_panel_parts = background_panel::build_background_panel(
         &window,
@@ -2324,363 +1996,33 @@ fn setup_editor_window_full(
     let root_overlay = Overlay::new();
     root_overlay.set_child(Some(&root));
 
-    // Transparent full-width strip over the checkerboard: drag the window + host toolbar.
-    let top_chrome = GtkBox::new(Orientation::Horizontal, 0);
-    top_chrome.add_css_class("editor-top-chrome");
-    top_chrome.set_halign(Align::Fill);
-    top_chrome.set_valign(Align::Start);
-    top_chrome.set_hexpand(true);
-    top_chrome.set_vexpand(false);
-    top_chrome.set_size_request(-1, EDITOR_TOP_CHROME_HEIGHT);
-    top_chrome.set_can_target(true);
-
-    let top_chrome_left = GtkBox::new(Orientation::Horizontal, 0);
-    top_chrome_left.set_hexpand(true);
-    let top_chrome_right = GtkBox::new(Orientation::Horizontal, 0);
-    top_chrome_right.set_hexpand(true);
-
-    toolbar.set_halign(Align::Center);
-    toolbar.set_valign(Align::Center);
-    toolbar.set_hexpand(false);
-    top_chrome.append(&top_chrome_left);
-    top_chrome.append(&toolbar);
-    top_chrome.append(&top_chrome_right);
-    canvas_with_toolbar.add_overlay(&top_chrome);
-
-    let zoom_control = GtkBox::new(Orientation::Horizontal, 0);
-    zoom_control.add_css_class("editor-floating-zoom");
-    zoom_control.set_halign(Align::Start);
-    zoom_control.set_valign(Align::End);
-    zoom_control.set_margin_start(16);
-    zoom_control.set_margin_bottom(16);
-    zoom_control.append(&zoom_minus_btn);
-    zoom_control.append(&zoom_button);
-    zoom_control.append(&zoom_plus_btn);
-    canvas_with_toolbar.add_overlay(&zoom_control);
-
-    let history_control = GtkBox::new(Orientation::Horizontal, 0);
-    history_control.add_css_class("editor-floating-history");
-    history_control.set_halign(Align::End);
-    history_control.set_valign(Align::End);
-    history_control.set_margin_end(16);
-    history_control.set_margin_bottom(16);
-    history_control.append(&history_group);
-    canvas_with_toolbar.add_overlay(&history_control);
-    canvas_with_toolbar.add_overlay(&zoom_popup);
+    chrome::install_window_chrome(chrome::WindowChromeInputs {
+        canvas_with_toolbar: &canvas_with_toolbar,
+        root_overlay: &root_overlay,
+        window: &window,
+        toolbar: &toolbar,
+        zoom_minus_btn: &zoom_minus_btn,
+        zoom_button: &zoom_button,
+        zoom_plus_btn: &zoom_plus_btn,
+        history_group: &history_group,
+        zoom_popup: &zoom_popup,
+    });
 
     if empty_drop_zone {
-        // Reuse the video editor's button / banner styles so the empty
-        // states look identical.
-        crate::recording::editor::ui_support::install_recording_editor_css();
-
-        let drop_center = GtkBox::new(Orientation::Vertical, 14);
-        drop_center.set_halign(Align::Center);
-        drop_center.set_valign(Align::Center);
-        drop_center.set_can_target(true);
-
-        let drop_icon = Image::from_icon_name("image-x-generic-symbolic");
-        drop_icon.add_css_class("recording-editor-empty-icon");
-        drop_icon.set_pixel_size(42);
-        drop_icon.set_halign(Align::Center);
-
-        let drop_title = Label::new(Some("Drop an image here"));
-        drop_title.add_css_class("recording-editor-empty-title");
-        drop_title.set_halign(Align::Center);
-
-        let drop_hint = Label::new(Some("PNG, JPEG, or WebP"));
-        drop_hint.add_css_class("recording-editor-empty-hint");
-        drop_hint.set_halign(Align::Center);
-
-        let open_btn = Button::with_label("Open Folder");
-        open_btn.set_has_frame(false);
-        open_btn.add_css_class("recording-editor-primary-button");
-        open_btn.add_css_class("recording-editor-empty-open-button");
-        open_btn.set_halign(Align::Center);
-
-        drop_center.append(&drop_icon);
-        drop_center.append(&drop_title);
-        drop_center.append(&drop_hint);
-        drop_center.append(&open_btn);
-
-        // Keep the empty state centered in the canvas pane, excluding the inspector.
-        canvas_with_toolbar.add_overlay(&drop_center);
-
-        // Loading banner (same as the video editor's "Loading video…").
-        let loading_box = GtkBox::new(Orientation::Horizontal, 8);
-        loading_box.add_css_class("recording-editor-drop-banner");
-        loading_box.set_can_target(false);
-        loading_box.set_halign(Align::Center);
-        let loading_spinner = Spinner::new();
-        loading_spinner.set_size_request(16, 16);
-        loading_spinner.set_can_target(false);
-        let loading_label = Label::new(Some("Loading image…"));
-        loading_label.add_css_class("recording-editor-drop-label");
-        loading_label.set_can_target(false);
-        loading_box.append(&loading_spinner);
-        loading_box.append(&loading_label);
-        let loading_revealer = Revealer::new();
-        loading_revealer.set_can_target(false);
-        loading_revealer.set_halign(Align::Center);
-        loading_revealer.set_valign(Align::Start);
-        loading_revealer.set_child(Some(&loading_box));
-        loading_revealer.set_transition_type(gtk4::RevealerTransitionType::SlideDown);
-        loading_revealer.set_reveal_child(false);
-        root_overlay.add_overlay(&loading_revealer);
-
-        let loading = Rc::new(Cell::new(false));
-
-        let app_open = app.clone();
-        let window_open = window.clone();
-        let revealer_open = loading_revealer.clone();
-        let spinner_open = loading_spinner.clone();
-        let loading_open = loading.clone();
-        let open_btn_ref = open_btn.clone();
-        open_btn.connect_clicked(move |_| {
-            show_open_image_dialog(
-                &app_open,
-                &window_open,
-                &revealer_open,
-                &spinner_open,
-                loading_open.clone(),
-                &open_btn_ref,
-            );
-        });
-
-        let drop_target = DropTarget::new(gio::File::static_type(), gdk::DragAction::COPY);
-        let app_drop = app.clone();
-        let window_drop = window.clone();
-        let revealer_drop = loading_revealer.clone();
-        let spinner_drop = loading_spinner.clone();
-        let loading_drop = loading.clone();
-        let open_btn_drop = open_btn.clone();
-        drop_target.connect_drop(move |_, value, _x, _y| {
-            if loading_drop.get() {
-                return false;
-            }
-            let Ok(file) = value.get::<gio::File>() else {
-                return false;
-            };
-            let Some(path) = file.path() else {
-                return false;
-            };
-            if !is_supported_image_path(&path) {
-                return false;
-            }
-            // Load into the SAME window (no close/reopen), like the video
-            // editor does.
-            load_image_into_editor(
-                &app_drop,
-                &window_drop,
-                path.to_path_buf(),
-                &revealer_drop,
-                &spinner_drop,
-                loading_drop.clone(),
-                &open_btn_drop,
-            );
-            true
-        });
-        root_overlay.add_controller(drop_target);
+        empty_state::install_empty_drop_zone(app, &window, &canvas_with_toolbar, &root_overlay);
     }
 
     window.set_child(Some(&root_overlay));
 
-    // Full-width canvas chrome + top inspector band move the window; tools/canvas stay interactive.
-    install_window_drag(&top_chrome, &window);
-    install_top_bar_window_drag(&root_overlay, &window);
-    install_edge_resize(&root_overlay, &window);
-
-    let update_canvas_content_size: Rc<dyn Fn()> = Rc::new({
-        let state = state.clone();
-        let zoom_level = zoom_level.clone();
-        let zoom_label = zoom_label.clone();
-        let zoom_header_label = zoom_header_label.clone();
-        let drawing_area = drawing_area.clone();
-        let _canvas_overlay = canvas_overlay.clone();
-        let canvas_scroller = canvas_scroller.clone();
-        let _window = window.downgrade();
-        move || {
-            let (
-                image_w,
-                image_h,
-                background_style,
-                background_padding,
-                background_insert,
-                background_aspect_ratio,
-                has_background,
-                crop_rect,
-                crop_mode_active,
-            ) = {
-                let st = state.lock().unwrap();
-                (
-                    st.working_image.width().max(1) as i32,
-                    st.working_image.height().max(1) as i32,
-                    st.background_style.clone(),
-                    st.background_padding,
-                    st.background_insert,
-                    st.background_aspect_ratio,
-                    st.background_style != BackgroundStyle::None,
-                    st.draft_crop_rect().or(st.crop_selection),
-                    st.selected_tool == Tool::Crop,
-                )
-            };
-
-            let mut virtual_w = image_w as f64;
-            let mut virtual_h = image_h as f64;
-
-            if has_background {
-                let layout = BackgroundComposition::new(virtual_w, virtual_h)
-                    .with_style(background_style)
-                    .with_padding(background_padding)
-                    .with_insert(background_insert)
-                    .with_alignment(BackgroundAlignment::Center)
-                    .with_corner_radius(18.0)
-                    .with_aspect_ratio(background_aspect_ratio)
-                    .compute();
-                virtual_w = layout.canvas_width;
-                virtual_h = layout.canvas_height;
-            }
-
-            let scroller_width = canvas_scroller.allocated_width().max(1) as f64;
-            let scroller_height = canvas_scroller.allocated_height().max(1) as f64;
-            // Keep fit-to-view math below the floating toolbar strip.
-            let top_inset = canvas_padding + EDITOR_TOP_CHROME_HEIGHT;
-            let available_width = (scroller_width - (canvas_padding * 2 + 2) as f64).max(1.0);
-            let available_height =
-                (scroller_height - (top_inset + canvas_padding + 2) as f64).max(1.0);
-
-            // Use the minimum of width and height to maintain aspect ratio and prevent asymmetric growth
-            let available_size = available_width.min(available_height);
-
-            // Layout scale without zoom - used for content size (prevents window from growing on zoom)
-            let layout_scale = (available_size / virtual_w.min(virtual_h)).min(1.0_f64);
-            // Rendering scale includes zoom for visual display
-            let scale = layout_scale * zoom_level.get().max(0.1_f64);
-
-            let fitted_w = (virtual_w * scale).round().max(1.0) as i32;
-            let fitted_h = (virtual_h * scale).round().max(1.0) as i32;
-
-            let (overflow_left, overflow_top, overflow_right, overflow_bottom) = if has_background {
-                (0.0, 0.0, 0.0, 0.0)
-            } else {
-                canvas::crop_canvas_overflow(
-                    crop_rect,
-                    image_w as f64,
-                    image_h as f64,
-                    scale,
-                    crop_mode_active,
-                )
-            };
-
-            let canvas_w = fitted_w
-                + canvas_padding * 2
-                + overflow_left.round() as i32
-                + overflow_right.round() as i32;
-            // Extra top inset so zoomed content cannot sit under the toolbar.
-            let canvas_h = fitted_h
-                + top_inset
-                + canvas_padding
-                + overflow_top.round() as i32
-                + overflow_bottom.round() as i32;
-
-            drawing_area.set_content_width(canvas_w);
-            drawing_area.set_content_height(canvas_h);
-            let percent_str = format!("{}%", (scale * 100.0).round().max(1.0) as i32);
-            zoom_label.set_label(&percent_str);
-            zoom_header_label.set_label(&percent_str);
-        }
-    });
-    update_canvas_content_size();
-
-    {
-        let update_canvas_content_size_tick = update_canvas_content_size.clone();
-        let state_canvas_tick = state.clone();
-        let zoom_level_tick = zoom_level.clone();
-        // Signature tracks the quantities that actually change the *visible* canvas size.
-        // Crucially, raw crop-rect coordinates are NOT included here.  Instead we compute
-        // the capped overflow bucket that crop_canvas_overflow() would return and store
-        // only that.  Because the function caps every side to 180 px, the bucket stays
-        // constant throughout an outside-image drag gesture — no relayout churn occurs.
-        let last_canvas_signature = Rc::new(Cell::new([
-            0_i32, // scroller width
-            0_i32, // scroller height
-            0_i32, // image width
-            0_i32, // image height
-            0_i32, // overflow left (px, capped)
-            0_i32, // overflow top  (px, capped)
-            0_i32, // overflow right (px, capped)
-            0_i32, // overflow bottom (px, capped)
-            0_i32, // crop mode active
-            0_i32, // zoom percentage
-            0_i32, // background enabled
-            0_i32, // background padding (tenths)
-            0_i32, // background insert (tenths)
-            0_i32, // background aspect ratio
-        ]));
-        let last_canvas_signature_tick = last_canvas_signature.clone();
-        canvas_scroller.add_tick_callback(move |scroller, _| {
-            let width = scroller.allocated_width();
-            let height = scroller.allocated_height();
-            let signature = {
-                let st = state_canvas_tick.lock().unwrap();
-                let img_w = st.working_image.width().max(1) as i32;
-                let img_h = st.working_image.height().max(1) as i32;
-                let crop_mode_active = st.selected_tool == Tool::Crop;
-                let crop_rect = st.draft_crop_rect().or(st.crop_selection);
-                let has_background = st.background_style != BackgroundStyle::None;
-                let background_padding = (st.background_padding * 10.0).round() as i32;
-                let background_insert = (st.background_insert * 10.0).round() as i32;
-                let background_aspect_ratio = st.background_aspect_ratio as i32;
-                let zoom_percentage = (zoom_level_tick.get() * 100.0_f64).round() as i32;
-
-                // Compute the same scale the layout function uses so we get the
-                // same overflow values without duplicating the full layout calculation.
-                let virtual_w = img_w as f64;
-                let virtual_h = img_h as f64;
-                let top_inset = canvas_padding + EDITOR_TOP_CHROME_HEIGHT;
-                let available_w = (width as f64 - (canvas_padding * 2 + 2) as f64).max(1.0);
-                let available_h =
-                    (height as f64 - (top_inset + canvas_padding + 2) as f64).max(1.0);
-
-                let available_size = available_w.min(available_h);
-                let layout_scale = (available_size / virtual_w.min(virtual_h)).min(1.0_f64);
-                let _scale = layout_scale * zoom_level_tick.get().max(0.1_f64);
-
-                let (ol, ot, or_, ob) = if has_background {
-                    (0.0, 0.0, 0.0, 0.0)
-                } else {
-                    canvas::crop_canvas_overflow(
-                        crop_rect,
-                        img_w as f64,
-                        img_h as f64,
-                        layout_scale,
-                        crop_mode_active,
-                    )
-                };
-
-                [
-                    width,
-                    height,
-                    img_w,
-                    img_h,
-                    ol.round() as i32,
-                    ot.round() as i32,
-                    or_.round() as i32,
-                    ob.round() as i32,
-                    if crop_mode_active { 1 } else { 0 },
-                    zoom_percentage,
-                    if has_background { 1 } else { 0 },
-                    background_padding,
-                    background_insert,
-                    background_aspect_ratio,
-                ]
-            };
-            if width > 0 && signature != last_canvas_signature_tick.get() {
-                last_canvas_signature_tick.set(signature);
-                update_canvas_content_size_tick();
-            }
-            glib::ControlFlow::Continue
-        });
-    }
+    let update_canvas_content_size = canvas_layout::install_canvas_layout(
+        &state,
+        &drawing_area,
+        &canvas_scroller,
+        &zoom_level,
+        &zoom_label,
+        &zoom_header_label,
+        canvas_padding,
+    );
 
     // Eyedropper
     color_picker::connect_eyedropper_activation(
@@ -3675,9 +3017,47 @@ mod tests {
 
     fn production_editor_window_source() -> String {
         let mod_src = include_str!("mod.rs");
-        let inspectors_src = include_str!("inspectors.rs");
+        let inspectors_mod = include_str!("inspectors/mod.rs");
+        let inspectors_shell = include_str!("inspectors/shell.rs");
+        let inspectors_select = include_str!("inspectors/select.rs");
+        let inspectors_crop = include_str!("inspectors/crop.rs");
+        let inspectors_stroke = include_str!("inspectors/stroke.rs");
+        let inspectors_text = include_str!("inspectors/text.rs");
+        let inspectors_number = include_str!("inspectors/number.rs");
+        let inspectors_obfuscate = include_str!("inspectors/obfuscate.rs");
+        let background_assets_src = include_str!("background_assets.rs");
+        let canvas_layout_src = include_str!("canvas_layout.rs");
+        let effects_src = include_str!("effects.rs");
+        let chrome_src = include_str!("chrome.rs");
+        let empty_state_src = include_str!("empty_state.rs");
         let mod_prod = mod_src.split("#[cfg(test)]").next().unwrap_or(mod_src);
-        format!("{mod_prod}\n{inspectors_src}")
+        let inspectors_prod = inspectors_mod
+            .split("#[cfg(test)]")
+            .next()
+            .unwrap_or(inspectors_mod);
+        let background_assets_prod = background_assets_src
+            .split("#[cfg(test)]")
+            .next()
+            .unwrap_or(background_assets_src);
+        let canvas_layout_prod = canvas_layout_src
+            .split("#[cfg(test)]")
+            .next()
+            .unwrap_or(canvas_layout_src);
+        let effects_prod = effects_src
+            .split("#[cfg(test)]")
+            .next()
+            .unwrap_or(effects_src);
+        let chrome_prod = chrome_src
+            .split("#[cfg(test)]")
+            .next()
+            .unwrap_or(chrome_src);
+        let empty_state_prod = empty_state_src
+            .split("#[cfg(test)]")
+            .next()
+            .unwrap_or(empty_state_src);
+        format!(
+            "{mod_prod}\n{inspectors_prod}\n{inspectors_shell}\n{inspectors_select}\n{inspectors_crop}\n{inspectors_stroke}\n{inspectors_text}\n{inspectors_number}\n{inspectors_obfuscate}\n{background_assets_prod}\n{canvas_layout_prod}\n{effects_prod}\n{chrome_prod}\n{empty_state_prod}"
+        )
     }
 
     #[test]
@@ -3721,12 +3101,14 @@ mod tests {
 
     #[test]
     fn editor_uses_full_top_bar_for_window_dragging() {
-        let source = include_str!("mod.rs");
-        let production_source = source.split("#[cfg(test)]").next().unwrap_or(source);
+        let production_source = production_editor_window_source();
         assert!(
-            production_source.contains("install_window_drag(&top_chrome, &window);")
-                && production_source
+            (production_source.contains("install_window_drag(&top_chrome, &window);")
+                || production_source.contains("install_window_drag(&top_chrome, window);"))
+                && (production_source
                     .contains("install_top_bar_window_drag(&root_overlay, &window);")
+                    || production_source
+                        .contains("install_top_bar_window_drag(root_overlay, window);"))
                 && production_source.contains("editor-top-chrome")
                 && production_source.contains("canvas_with_toolbar.add_overlay(&top_chrome);")
                 && production_source.contains("EDITOR_TOP_CHROME_HEIGHT")
